+++ packages/farms/constants/priceHelperLps/1.ts
@@ -1,4 +1,4 @@
-import { SerializedFarmConfig } from '../../types'
+import { SerializedFarmConfig } from '@pancakeswap/farms'
 
 const priceHelperLps: SerializedFarmConfig[] = []
