+++ packages/farms/constants/priceHelperLps/97.ts
@@ -1,0 +1,5 @@
+import { SerializedFarmConfig } from '@pancakeswap/farms'
+
+const priceHelperLps: SerializedFarmConfig[] = []
+
+export default priceHelperLps
