--- src/config/constants/priceHelperLps/farms/1.ts
@@ -1,9 +1,0 @@
-// import { serializeTokens } from 'utils/serializeTokens'
-import { SerializedFarmConfig } from '../../types'
-// import { ethereumTokens } from '../tokens'
-
-// const serializedTokens = serializeTokens(ethereumTokens)
-
-const priceHelperLps: SerializedFarmConfig[] = []
-
-export default priceHelperLps
