--- src/config/constants/priceHelperLps/farms/97.ts
@@ -1,6 +1,0 @@
-// import { bscTestnetTokens } from '@pancakeswap/tokens'
-import { SerializedFarmConfig } from '../../types'
-
-const priceHelperLps: SerializedFarmConfig[] = []
-
-export default priceHelperLps
