+++ src/config/constants/priceHelperLps/index.ts
@@ -1,27 +1,11 @@
+import { getFarmsPriceHelperLpFiles } from '@pancakeswap/farm-constants/priceHelperLps/getFarmsPriceHelperLpFiles'
 import { ChainId } from '@pancakeswap/sdk'
-import FarmsBscPriceHelper from './farms/56'
-import FarmsBscTestnetPriceHelper from './farms/97'
-import FarmsEthereumPriceHelper from './farms/1'
-import FarmsGoerliPriceHelper from './farms/5'
-import PoolsBscPriceHelper from './pools/56'
-import PoolsBscTestnetPriceHelper from './pools/97'
 import PoolsEthereumPriceHelper from './pools/1'
 import PoolsGoerliPriceHelper from './pools/5'
+import PoolsBscPriceHelper from './pools/56'
+import PoolsBscTestnetPriceHelper from './pools/97'
 
-export const getFarmsPriceHelperLpFiles = (chainId: ChainId) => {
-  switch (chainId) {
-    case ChainId.BSC:
-      return FarmsBscPriceHelper
-    case ChainId.BSC_TESTNET:
-      return FarmsBscTestnetPriceHelper
-    case ChainId.ETHEREUM:
-      return FarmsEthereumPriceHelper
-    case ChainId.GOERLI:
-      return FarmsGoerliPriceHelper
-    default:
-      return []
-  }
-}
+export { getFarmsPriceHelperLpFiles }
 
 export const getPoolsPriceHelperLpFiles = (chainId: ChainId) => {
   switch (chainId) {
