+++ src/state/farms/index.ts
@@ -1,5 +1,5 @@
 import { formatEther } from '@ethersproject/units'
-import { getFarmConfig } from '@pancakeswap/farm-constants'
+import { getFarmConfig } from '@pancakeswap/farms/constants'
 import { createFarmFetcher } from '@pancakeswap/farms'
 import { ChainId } from '@pancakeswap/sdk'
 import { createAsyncThunk, createSlice, isAnyOf } from '@reduxjs/toolkit'
