+++ src/views/Migration/hook/V1/Pool/useFetchPublicPoolsData.ts
@@ -1,6 +1,6 @@
 import { ChainId } from '@pancakeswap/sdk'
 import { useAppDispatch } from 'state'
-import { getFarmConfig } from '@pancakeswap/farm-constants'
+import { getFarmConfig } from '@pancakeswap/farms/constants'
 import { useSlowRefreshEffect } from 'hooks/useRefreshEffect'
 import { fetchFarmsPublicDataAsync } from 'state/farmsV1/index'
